+++ CHANGES.md
@@ -6,12 +6,13 @@ Features:
 
 Improvements:
  - Persist active tab between sessions (#503)
+ - Do not upload file to big for the homeserver (#587)
 
 Other changes:
  -
 
 Bugfix:
- -
+ - Fix issue on upload error in loop (#587)
 
 Translations:
  -